+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Changed
 Removed
 
 Fixed
+ - unique_containers can properly deal with WellGroups
 
 ## v2.1.3 - 2016-04-07
 ---